refactor : periodTrade dto 디렉토리 수정
