+++ src/main/java/com/barter/domain/trade/periodtrade/dto/request/SuggestedPeriodTradeReqDto.java
@@ -1,4 +1,4 @@
-package com.barter.domain.trade.periodtrade.dto;
+package com.barter.domain.trade.periodtrade.dto.request;
 
 import java.util.List;
