+++ src/main/java/com/barter/domain/trade/periodtrade/dto/response/UpdatePeriodTradeResDto.java
@@ -1,4 +1,4 @@
-package com.barter.domain.trade.periodtrade.dto;
+package com.barter.domain.trade.periodtrade.dto.response;
 
 import java.time.LocalDateTime;
